# public/openapi.yaml
openapi: 3.0.3
info:
  title: Promptly API
  description: REST API для маркетплейсу AI-промптів (Next.js, Firebase, Stripe).
  version: 1.0.0
servers:
  - url: /
    description: Поточний origin (той самий, що й фронт)
...
paths:
...
  /api/purchase:
    post:
      tags: [Purchase]
      summary: Покупка промпту(ів) за кредитами
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  required: [promptId]
                  properties:
                    promptId: { type: string }
                - type: object
                  required: [type, promptIds]
                  properties:
                    type: { type: string, enum: [cart] }
                    promptIds: { type: array, items: { type: string } }
      responses:
        '200':
          description: Успішно
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
        '400':
          description: Невірний body або недостатньо кредитів
        '401':
          description: Не авторизовано (Bearer token)
...
  /api/search-bar-backgrounds/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Search bar backgrounds]
    patch:
      tags: [Search bar backgrounds]
    delete:
      tags: [Search bar backgrounds]
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
...